Revert "move compilePropertyAccessor to utils"
This reverts commit f7323f43b74549233dbf2460fbf6536ca6fa6feb.
